Rename ContactInfoBox to ContactCard to match Figma component name.
