+++ src/lib/components/ContactCard/ContactCard.scss
