--- src/lib/components/ContactInfoBox/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./ContactInfoBox.svelte";
