+++ src/stories/ContactCard.stories.ts
@@ -1,12 +1,12 @@
 import type { Meta, StoryObj } from "@storybook/svelte";
 
-import ContactInfoBox from "$lib/components/ContactInfoBox";
+import ContactCard from "$lib/components/ContactCard";
 
 import OfficePageTestContent from "../routes/about/organization/[slug]/__tests__/OfficePageTestContent";
 
 const meta = {
-  title: "Components/ContactInfoBox",
-  component: ContactInfoBox,
+  title: "Components/ContactCard",
+  component: ContactCard,
   tags: ["autodocs"],
   parameters: {
     design: {
@@ -14,7 +14,7 @@ const meta = {
       url: "https://www.figma.com/file/oGKbyCnCRRdNzLYbiags93/LDAF-Component-Library-USWDS-3.0.2?type=design&node-id=2836-8572&t=S7mXyVT1E3IbZMMx-4",
     },
   },
-} satisfies Meta<ContactInfoBox>;
+} satisfies Meta<ContactCard>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
